Some more associated cleanup.
